Change timeout argument type to long

+++ src/ru/nsu/ccfit/bogush/net/tou/segment/TOUSegment.java
@@ -23,7 +23,7 @@ public class TOUSegment extends TCPSegment {
         return dst;
     }
 
-    public void setTimeout(int timeout) {
+    public void setTimeout(long timeout) {
         this.timeout = System.currentTimeMillis() + timeout;
     }
